fix refactor of execFile() & its helper funcs
Now execFile() seems to work properly, and ncurses mode looks functional
except for box-drawing.
Error was due to indexing with a single variable 'i' in onlyCode(),
fixed by using 'i' to index original string, and 'j' to index output
string.

+++ src/kler.h
@@ -14,7 +14,7 @@
 #define DELAY 100000
 #define PADDING 2
 
-#define TAPE_POS (tape.tc - tape.start)
+#define TAPE_POS (unsigned int) (tape.tc - tape.start)
 
 enum {
 	NO_ERROR,
